add admin block and updateRole

+++ src/handlers/user/block.ts
@@ -6,13 +6,20 @@ import FollowModel from '../../models/follow.model';
 import UserModel from '../../models/user.model';
 
 // route:   POST /:userId/block
-// access:  logged-user
+// access:  logged-user / admin
 export const block = asyncHandler(async (req, res, next) => {
   // user cannot block himself
   if (res.locals.userId === req.params.userId)
     return next(new ApiError('cannot block your self', 400));
   const targetUser = await UserModel.findByPk(req.params.userId, { attributes: ['id'] });
   if (!targetUser) return next(new ApiError('user not found', 400));
+  // if logged-user is admin
+  const loggedUser = await UserModel.findByPk(res.locals.userId, { attributes: ['isAdmin'] });
+  if (loggedUser?.isAdmin) {
+    await targetUser.update({ isBlocked: true });
+    res.sendStatus(200);
+    return;
+  }
   // remove target user from follow if exists
   await FollowModel.destroy({ where: { userId: res.locals.userId, followId: targetUser.id } });
   // assert logged user not unblocked target user
